move calmdynamorecord to common
