+++ common/src/main/scala/uk/ac/wellcome/models/CalmDynamoRecord.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.transformer.models
+package uk.ac.wellcome.models
 
 import uk.ac.wellcome.utils.JsonUtil
 import scala.util.Try
